Renamed connection.py to _connection.py
